Fix working directory problem with GitHub Actions
I forgot to change the `working-directory` for all the tests in the new GitHub Actions, which broke things when they ran on GitHub.

I don't really need the `working-directory` entry here since we run the tests from the top-level directory in this lab, but I kept it because it helps make it clear that we're running the tests in the top-level directory.